Build in poetry; fix localdocker-run
Use the poetry base image.  Unfortunately, Dockerfile `COPY --from=FOO`
does not allow interpolation for `FOO` (at least, not in 18.02.0-ce).
And since there's no /bin/sh we can't import it as a builder stage,
because Docker seems to try to invoke something even with no steps?

So the poetry image is hard-coded for the actual build-step.  This is a
bug/limitation.

To test this, I tried actually using the simplified localdocker form of
this repo and realized that I was missing easy ways to find what to talk
to, so `make localdocker-run` no longer leaves the image attached but
instead returns, after showing some basic information.

In an ideal work, there would be a mode for Docker when using `-P` or
other exposure flags to print the exposed ports to stderr at start-up,
so this make target could remain attached sanely, instead of having to
reattach.