Optimize CI: remove caching, smaller deployer image
We don't need/want caching for our use-case, but I did want to show it
working and how it interacts with multi-stage Docker builds, so I
debugged it and now it's commented out as an example.

We don't need the custom builder image for deploys, and it's large,
so skip it.  At this point, with no tests running on the outer-Docker
image, the only thing we do rely upon it for is 'dep'.