cleanup init of poetry, fix dir bug
The poetry code noted that it was overly complex.  This was an
understatement and a sign of doing stupid things when tired.

Registering the web-server handlers at init() time meant that flag value
overrides on the command-line don't take effect.  The deferrals on
guarding existence didn't affect the time at which the web handling
objects were created.

Move the logic to a setup to be called by main go-routine after
flag-parsing.  Now we "guard existence" by just not registering the
handlers at all, which makes more sense.  It does mean that the map of
available first-level pages mutates after main is started, but that
should be fine for us.  If we get more complex, then we can switch to a
sync.Map if there might be other things happening concurrently, but we
setup the poetry handler before the webserver is started.

By reworking, we get rid of the sync/atomic nonsense around
double-invocation, and all that smell.  We break up into "bit which does
the work and just returns an error, but has no logging" and "logging
wrapper which handles whether logger is nil or not".

Left the GuardExistence handling in for the root page renderer, in case
something else wants it.