Time in µs as string; remove zombie log directive
First thing I checked was "format of the duration", but I didn't check
in JSON mode and the formatting differs between the two.  So I saw a
nice string with a µs suffix, while JSON just recorded an integer which
happens to be Nanoseconds (because Go, but non-obvious to others, and
not a useful scale).

Switch field from `duration` key to `duration.µs` key, and records as a
string, to two decimal places.  We always record in µs so that units
don't have to be parsed (and the unit is _only_ in the fieldname, not
the value).

I accidentally left the old logging defer in.  This nukes it, new
response log only.