Tag make-built binary with build timestamp
Update version handling to show the date, but only if provided.

Fix quoting for Go arguments so that linker values containing whitespace
will make it through intact to `go tool` stage.  Notably, the shell
command-line wraps the `-ldflags` parameter in single-quotes, so that as
long as the string to be embedded doesn't contain those, we're good;
then use double-quotes for the individual parameters, where they're
only affecting boundaries, not presence/absence of the
outer-sq-inhibited interpolation.