Log field changes, backwards-incompatible
I knew that using `.` in a log-field name was iffy, even if it made a
kind of hierarchical sense without having to bother with actual
hierarchy.  I didn't think about ease of parsing with `jq`, and use of
`.fieldname` instead of `.["fieldname"]`.

Replace `.` with `_` in log field keys.
Similarly replace `µs` with `us` for the duration: Unicode is fine per
JSON, but `jq` dislikes it unquoted so it makes life more awkward.  It
might be "more correct" but it's less convenient to _use_.

Changed the 404 handler's log to include the status code as a field;
note that now that we're using `httpsnoop` this isn't needed and
actually results in an extra line, but I want to keep in the example of
pulling the logger from the `context`, so I'm leaving it in.

Have the `internal/version` package provide a function to return a list
of key/value pairs for logging and let it take care of all logic around
versioning and which fields might or might not exist, rather than
propagate build-time "possible existence" into main.  Cleans up the
location env-var handling too.