Unbreak Docker ARG overrides, interpolate poetry image
Two key points I had somehow not grokked when reading the Dockerfile
reference documentation:
1. each `FROM` will only interpolate non-default values for an `ARG` if
   the `ARG` is declared before the _first_ `FROM`, not _this_ `FROM`
2. as long as you declare the `ARG` before the first `FROM`, you _don’t_
   need to repeat its value when re-declaring it

Thus an override for the `scratch` was not working (I'd only ever tested
the builder image) and all the repeated `ARG` values were unnecessary.

Further, there's a way to `COPY --from` an ARG value image name; the
issue for Docker is that they want to build a complete graph of the
images which need to be pulled before they start to pull, which is fair,
while disallowing `ENV` for one interpolation _within_ a `FROM` block
opens up a new world of inconsistency, so they don't.  But instead, we
can have an _empty_ stage which just pulls the image and names it to a
static reference, as described by @tonistiigi in
<https://github.com/moby/moby/issues/34482#issuecomment-332298635>

Changes:
* Moved all the `ARG` declarations, with defaults, up to the top of
  `build/Dockerfile` and replace with declaration of usage in-line
* Make POETRY_IMAGE interpolation work right, using an extra (empty)
  stage to pull the image
* Add the Predefined ARGs to the set grabbed and munged by `GNUmakefile`
  when constructing the `docker build` command-line; move the complete
  list into a standalone variable, so that the `print-%` rule can be
  used with it; you can now do `make print-AVAILABLE_DOCKER_ARGS` to
  see them all
* Add to README a "Build Tricks" section with just one trick so far,
  showing how to pass through an `http_proxy` (which ... isn't used
  by this setup, currently, but proves the point, which is the whole
  point of the dummyapp)
* Bump the default version for `go build` builds to `0.0.4`