Fix image tagging Golang version
Commit b14dd5c58 unbroke `ARG` overrides in the `Dockerfile`, which
meant that an attempt to change the image unbroke and the bit of
`.circleci/config.yml` which overrode the Builder Image came into
effect.

Fix the documentation which claimed otherwise, as I'd documented what
I'd observed.

The CI image I'm using is code-named without the Go version in it, and
had switched to Go 1.10.  Unfortunately the `docker build` system can't
use interrogation commands inside the Docker image to populate `LABEL`
values, so we have to pass data statically, and the version override
being passed in still claimed 1.9.4.  The only effect was mis-labelling
the metadata on the image.  But fix it.