make: move setup to .INTERMEDIATE
I briefly was going to make it `.INTERMEDIATE` if `NODEP` defined, but
then realized that because this isn't an `-update`, we really don't want
the rebuild forced based on this.  If a previous build succeeded, we
have the deps, unless Gopkg.lock has been updated.

I reason that this is probably more correct.