make: more .INTERMEDIATE, more ordering
Using `native: setup $(BINNAME)`, I was seeing `$(BINNAME)` be invoked
before `setup`.  The "left-to-right if nothing else" was not protecting
us.  Also, this was fragile in the face of concurrent make.

So switch to using order-only dependencies, where the phony or
intermediate target lists the targets as before, but then each target
has an order-only dependency on those to the left of it.

As long as we don't ever have:

    foo: alpha beta
    beta: | alpha
    bar: beta alpha
    alpha: | beta

then we should be okay.  If we need more complex "deploy action"
pipelines, then Make is not the tool for us.