Wrap logging properly; style fixes
Using a type alias for the logger was a fast hack to get moving, but
is an approach which has both pros and cons.  Changed to use our own
shim which should compile down to almost identical space and probably
an extra function later in some calls, but which lets us be much more
explicit about the API we use.

Fixed some logging `Foof` calls to `Foo`, since with structured logging
we want all those variables in fields, not in the string.  With this,
the actual interface we export is much smaller than that of
`logrus.FieldLogger`.

Fix various golint issues, leaving just one which _should_ be style
violating because it's compile-time logic masking.