Nuke the GNUmakefile, use shell scripts
Make is superficially the right tool for the job, but all the
dependency finding is already handled by the Go tool, so what we really
wanted was "here's a bunch of tasks, and this super-task requires that
these other tasks be run in a specific order, where some of them are
optional".  That's not Make.  The dependency order is not guaranteed,
and trying to avoid spurious rebuilds led to complexity and fragility.

Add a common shell library to abstract away the details of how docker
tags vary depending upon the features, etc, and some small and
simple-ish scripts using that common library.  Ensure every script is
checked with shfmt/shellcheck.