Switch to no-entrypoint, bump version
Using the new `heroku labs:enable runtime-empty-entrypoint` support from
Heroku, we no longer need to set an entry-point and so can have an
entirely consistent image and invocation, whether building upon
`scratch` or `alpine` or any other reasonable base.

Make the switch, declare this to be worthy of a minor number bump, we
are now `v0.1.0`!

Bump deps too:
 * go.pennock.tech/hmetrics: gained new tests, got a version bump
 * golang.org/x/crypto: repo changes not affecting our packages
 * golang.org/x/sys: mostly portability fixes